엑티비티 디테일에서도 바로 수정가능하게

+++ src/components/activity/ActivityDetail.vue
@@ -491,14 +491,30 @@ export default {
         }
       } else if (param === 'modify') {
         if (window.confirm('수정을 승인하시겠습니까?')) {
-          this.item.host_key = this.item.Host.host_key
-          let originalAcitivityKey = this.item.category_two
-          this.$http.put(`/activity/front/${originalAcitivityKey}`, this.item)
-          .then(_ => this.$http.put(`/activity/approve/${originalAcitivityKey}`))
-          .then(r => {
-            this.$http.delete(`/activity/admin/${this.item.activity_key}`)
+          if (this.item.status === '9') {
+            this.item.host_key = this.item.Host.host_key
+            let originalAcitivityKey = this.item.category_two
+            this.$http.put(`/activity/front/${originalAcitivityKey}`, this.item)
+              .then(_ => this.$http.put(`/activity/approve/${originalAcitivityKey}`))
+              .then(r => {
+                this.$http.delete(`/activity/admin/${this.item.activity_key}`)
+                  .then(result => {
+                    this.item.status = 3
+                    window.alert('수정 승인 완료')
+                    this.$router.push('/approve')
+                  })
+                  .catch(e => {
+                    console.log(e)
+                    window.alert(moment().format() + '새로고침 후에도 작동이 안된다면 유성이에게 해당 화면을 보여주세요')
+                  })
+              })
+              .catch(e => {
+                console.log(e)
+                window.alert(moment().format() + '새로고침 후에도 작동이 안된다면 유성이에게 해당 화면을 보여주세요')
+              })
+          } else {
+            this.$http.put(`/activity/front/${this.item.activity_key}`, this.item)
               .then(result => {
-                this.item.status = 3
                 window.alert('수정 승인 완료')
                 this.$router.push('/approve')
               })
@@ -506,11 +522,7 @@ export default {
                 console.log(e)
                 window.alert(moment().format() + '새로고침 후에도 작동이 안된다면 유성이에게 해당 화면을 보여주세요')
               })
-          })
-          .catch(e => {
-            console.log(e)
-            window.alert(moment().format() + '새로고침 후에도 작동이 안된다면 유성이에게 해당 화면을 보여주세요')
-          })
+          }
         }
       } else if (param === 'delete') {
         if (window.confirm('위킨을 삭제 하시겠습니까?')) {
